refactor(subway): separate class location into different packages
