+++ src/main/java/subway/controller/path/QuickestPathController.java
@@ -2,7 +2,7 @@ package subway.controller.path;
 
 import java.util.List;
 import org.jgrapht.alg.shortestpath.DijkstraShortestPath;
-import subway.domain.TimeGraphRepository;
+import subway.domain.graph.TimeGraphRepository;
 import subway.view.InputView;
 import subway.view.OutputView;
