+++ src/main/java/subway/domain/line/Line.java
@@ -1,8 +1,9 @@
-package subway.domain;
+package subway.domain.line;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import subway.domain.station.Station;
 
 public class Line {
